+++ packages/typespec-rust/CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Omit a `Default` implementation for omitted client options types.
 * Fixed another case of colliding locals.
+* Only emit one constructor for `oauth2` when multiple flows are described.
 
 ## 0.32.0 (2025-12-11)